requirements.txt
absl-py==0.15.0
aclick==0.2.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
configparser==5.1.0
cycler==0.11.0
Cython==0.29.28
dill==0.3.4
dm-tree==0.1.6
docker-pycreds==0.4.0
flatbuffers==1.12
fonttools==4.29.0
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.41.0
httplib2==0.20.4
idna==3.3
joblib==1.1.1
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
matplotlib==3.5.1
mpmath==1.2.1
numpy>=1.21.6
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas>=1.3.5
pathtools==0.1.2
Pillow==9.0.0
portalocker==2.4.0
promise==2.3
protobuf==3.19.0
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
PyYAML
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==2.0.0
scipy
sentencepiece==0.1.96
sentry-sdk==1.4.3
seqeval==1.2.2
shortuuid==1.0.7
six==1.15.0
sklearn==0.0
scikit-learn==1.2.0
smmap==5.0.0
subprocess32==3.5.4
sympy==1.9
tabulate==0.8.9
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-addons
tensorflow-datasets==4.5.2
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem
tensorflow-metadata==1.7.0
tensorflow-model-optimization==0.7.2
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tqdm==4.62.3
typeguard==2.13.2
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
wandb==0.12.6
Werkzeug==2.0.2
wrapt==1.12.1
yaspin==2.1.0